No float16

+++ pix2pix-cpu.py
@@ -3,7 +3,7 @@ import torch
 from diffusers import StableDiffusionInstructPix2PixPipeline, EulerAncestralDiscreteScheduler
 
 model_id = "timbrooks/instruct-pix2pix"
-pipe = StableDiffusionInstructPix2PixPipeline.from_pretrained(model_id, torch_dtype=torch.float16, safety_checker=None)
+pipe = StableDiffusionInstructPix2PixPipeline.from_pretrained(model_id, safety_checker=None)
 pipe.to("cpu")
 pipe.scheduler = EulerAncestralDiscreteScheduler.from_config(pipe.scheduler.config)
